+++ playa/cli.py
@@ -25,8 +25,12 @@ following keys (but will probably contain more in the future):
     - `objid`: the object number
     - `genno`: the generation number
     - `type`: the type of object this is
-    - `repr`: an arbitrary string representation of the object, **do not
-        depend too closely on the contents of this as it will change**
+    - `obj`: a best-effort JSON serialization of the object's
+      metadata.  In the case of simple objects like strings,
+      dictionaries, or lists, this is the object itself.  Object
+      references are converted to a string representation of the form
+      "<ObjRef:OBJID>", while content streams are reprented by their
+      properties dictionary.
 
 Bucking the trend of the last 20 years towards horribly slow
 Click-addled CLIs with deeply nested subcommands, anything else is
@@ -65,10 +69,15 @@ reading order, and so we can actually really extract the text from it
 
     playa --text tagged-foo.pdf
 
+And finally yes you can also extract images (not necessarily useful
+since they are frequently tiled and/or composited):
+
+    playa --images foo.pdf --imgdir outdir
 """
 
 import argparse
 import functools
+import getpass
 import itertools
 import json
 import logging
@@ -77,11 +86,16 @@ from pathlib import Path
 from typing import Any, Deque, Dict, Iterable, Iterator, List, TextIO, Tuple, Union
 
 import playa
-from playa import Document, Page
-from playa.pdftypes import ContentStream, ObjRef, resolve1
-from playa.structure import Element, ContentObject as StructContentObject, ContentItem
+from playa import Document, Page, PDFPasswordIncorrect, asobj
+from playa.data.metadata import asobj_document
+from playa.data.content import Image
+from playa.page import ContentObject, TextObject, ImageObject
+from playa.pdftypes import ContentStream, ObjRef, resolve1, LITERALS_DCT_DECODE
+from playa.structure import ContentItem
+from playa.structure import ContentObject as StructContentObject
+from playa.structure import Element
 from playa.utils import decode_text
-from playa.worker import _deref_page, PageRef
+from playa.worker import PageRef, _deref_page
 
 LOG = logging.getLogger(__name__)
 
@@ -118,6 +132,16 @@ def make_argparse() -> argparse.ArgumentParser:
         help="Decode content streams into raw bytes",
     )
     parser.add_argument(
+        "--content-objects",
+        action="store_true",
+        help="Extract content objects as JSON",
+    )
+    parser.add_argument(
+        "--explode-text",
+        action="store_true",
+        help="Explode text objects into constituent glyphs (are you sure?)",
+    )
+    parser.add_argument(
         "-x",
         "--text-objects",
         action="store_true",
@@ -133,6 +157,21 @@ def make_argparse() -> argparse.ArgumentParser:
         help="Extract logical structure tree as JSON",
     )
     parser.add_argument(
+        "--outline",
+        action="store_true",
+        help="Extract document outline as JSON",
+    )
+    parser.add_argument(
+        "--images",
+        action="store_true",
+        help="Extract images as... whatever",
+    )
+    parser.add_argument(
+        "--imgdir",
+        type=Path,
+        help="Extract image files here (default is not to extract).",
+    )
+    parser.add_argument(
         "-o",
         "--outfile",
         help="File to write output (or - for standard output)",
@@ -145,6 +184,18 @@ def make_argparse() -> argparse.ArgumentParser:
         type=int,
         help="Maximum number of worker processes to use",
         default=1,
+    )
+    parser.add_argument(
+        "--password",
+        help="Password for an encrypted PDF.  If not supplied, "
+        "will be read from the console.",
+        type=str,
+        default="",
+    )
+    parser.add_argument(
+        "--non-interactive",
+        help="Do not attempt to read a password from the console.",
+        action="store_true",
     )
     parser.add_argument(
         "--debug",
@@ -199,43 +250,14 @@ def extract_catalog(doc: Document, args: argparse.Namespace) -> None:
         args.outfile,
         indent=2,
         ensure_ascii=False,
-        default=repr,
+        default=asobj,
     )
 
 
 def extract_metadata(doc: Document, args: argparse.Namespace) -> None:
     """Extract random metadata."""
-    stuff = {
-        "pdf_version": doc.pdf_version,
-        "is_printable": doc.is_printable,
-        "is_modifiable": doc.is_modifiable,
-        "is_extractable": doc.is_extractable,
-    }
-    pages = []
-    for page in doc.pages:
-        pages.append(
-            {
-                "objid": page.pageid,
-                "label": page.label,
-                "mediabox": page.mediabox,
-                "cropbox": page.cropbox,
-                "rotate": page.rotate,
-            }
-        )
-    stuff["pages"] = pages
-    objects = []
-    for obj in doc.objects:
-        # TODO: Add method to JSON serialize indirect objects
-        objects.append(
-            {
-                "objid": obj.objid,
-                "genno": obj.genno,
-                "type": type(obj.obj).__name__,
-                "repr": repr(obj.obj),
-            }
-        )
-    stuff["objects"] = objects
-    json.dump(stuff, args.outfile, indent=2, ensure_ascii=False, default=repr)
+    metadata = asobj_document(doc, exclude={"structure", "outline"})
+    json.dump(metadata, args.outfile, indent=2, ensure_ascii=False)
 
 
 def decode_page_spec(doc: Document, spec: str) -> Iterator[int]:
@@ -250,36 +272,21 @@ def decode_page_spec(doc: Document, spec: str) -> Iterator[int]:
         yield from pages
 
 
-def get_text_json(page: Page) -> List[str]:
+def flatten_harder(itor: Iterator[ContentObject]) -> Iterator[ContentObject]:
+    for obj in itor:
+        if isinstance(obj, TextObject):
+            yield from obj
+        else:
+            yield obj
+
+
+def get_text_json(page: Page, explode_text: bool = False) -> List[str]:
     objs = []
-    for text in page.texts:
-        tstate = text.textstate
-        # Prune these objects somewhat (FIXME: need a method that will
-        # serialize only non-default values and run _asdict as needed)
-        textstate = {
-            "line_matrix": tstate.line_matrix,
-            "fontsize": tstate.fontsize,
-            "render_mode": tstate.render_mode,
-        }
-        if tstate.font is not None:
-            textstate["font"] = {
-                "fontname": tstate.font.fontname,
-                "vertical": tstate.font.vertical,
-            }
-        gstate = {
-            "scs": text.gstate.scs._asdict(),
-            "scolor": text.gstate.scolor._asdict(),
-            "ncs": text.gstate.ncs._asdict(),
-            "ncolor": text.gstate.ncolor._asdict(),
-        }
-        obj = {
-            "chars": text.chars,
-            "bbox": text.bbox,
-            "textstate": textstate,
-            "gstate": gstate,
-            "mcstack": [mcs._asdict() for mcs in text.mcstack],
-        }
-        objs.append(json.dumps(obj, indent=2, ensure_ascii=False, default=repr))
+    itor = flatten_harder(page.texts) if explode_text else page.texts
+    for text in itor:
+        objs.append(
+            json.dumps(asobj(text), indent=2, ensure_ascii=False, default=asobj)
+        )
     return objs
 
 
@@ -288,9 +295,41 @@ def extract_text_objects(doc: Document, args: argparse.Namespace) -> None:
     pages = decode_page_spec(doc, args.pages)
     print("[", file=args.outfile)
     last = None
-    for obj in itertools.chain.from_iterable(doc.pages[pages].map(get_text_json)):
+    for obj in itertools.chain.from_iterable(
+        doc.pages[pages].map(
+            functools.partial(get_text_json, explode_text=args.explode_text)
+        )
+    ):
         if last is not None:
-            print(last, ",", sep="", file=args.outfile)
+            print(last, end=",\n", sep="", file=args.outfile)
+        last = obj
+    if last is not None:
+        print(last, file=args.outfile)
+    print("]", file=args.outfile)
+
+
+def get_content_json(page: Page, explode_text: bool = False) -> List[str]:
+    objs = []
+    itor = flatten_harder(page.flatten()) if explode_text else page.flatten()
+    for obj in itor:
+        objdict = asobj(obj)
+        objdict["object_type"] = obj.object_type
+        objs.append(json.dumps(objdict, indent=2, ensure_ascii=False, default=asobj))
+    return objs
+
+
+def extract_content_objects(doc: Document, args: argparse.Namespace) -> None:
+    """Extract content objects as JSON."""
+    pages = decode_page_spec(doc, args.pages)
+    print("[", file=args.outfile)
+    last = None
+    for obj in itertools.chain.from_iterable(
+        doc.pages[pages].map(
+            functools.partial(get_content_json, explode_text=args.explode_text)
+        )
+    ):
+        if last is not None:
+            print(last, end=",\n", sep="", file=args.outfile)
         last = obj
     if last is not None:
         print(last, file=args.outfile)
@@ -353,7 +392,7 @@ def _extract_content_object(
     kid: StructContentObject, indent: int, outfh: TextIO
 ) -> bool:
     ws = " " * indent
-    text = json.dumps(kid.props, default=repr)
+    text = json.dumps(kid.props, default=asobj)
     print(f"{ws}{text}", end="", file=outfh)
     return True
 
@@ -425,9 +464,64 @@ def extract_structure(doc: Document, args: argparse.Namespace) -> None:
     print("]", file=args.outfile)
 
 
-def main() -> None:
+def extract_outline(doc: Document, args: argparse.Namespace) -> None:
+    """Extract logical outline as JSON."""
+    if doc.outline is None:
+        LOG.info("Document has no logical structure")
+        print("{}", file=args.outfile)
+        return
+    metadata = asobj(doc.outline)
+    json.dump(metadata, args.outfile, indent=2, ensure_ascii=False)
+
+
+def get_images(page: Page, imgdir: Path) -> List[Tuple[Path, Image]]:
+    images = []
+    for idx, img in enumerate(page.flatten(ImageObject)):
+        if img.xobjid is None:
+            text_bbox = ",".join(str(round(x)) for x in img.bbox)
+            imgname = f"page{page.page_idx}-{idx}-inline-{text_bbox}"
+        else:
+            imgname = f"page{page.page_idx}-{idx}-{img.xobjid}"
+        imgpath = imgdir / imgname
+        # FIXME: only really support plain JPEG for now...
+        fp = img.stream.get_filters()
+        if fp:
+            filters, params = zip(*fp)
+            for f in filters:
+                if f in LITERALS_DCT_DECODE:
+                    imgpath = imgpath.with_suffix(".jpg")
+        if imgpath.suffix == "":
+            imgpath = imgpath.with_suffix(".dat")
+        with open(imgpath, "wb") as outfh:
+            outfh.write(img.stream.buffer)
+        images.append((imgpath, asobj(img)))
+    return images
+
+
+def extract_images(doc: Document, args: argparse.Namespace) -> None:
+    """Extract images."""
+    pages = decode_page_spec(doc, args.pages)
+    print("[", file=args.outfile, end="")
+    if args.imgdir is not None:
+        args.imgdir.mkdir(exist_ok=True, parents=True)
+    last = None
+    for page, images in enumerate(
+        doc.pages[pages].map(functools.partial(get_images, imgdir=args.imgdir))
+    ):
+        for path, image in images:
+            if last is not None:
+                print(last, end=",\n", sep="", file=args.outfile)
+            image["page_idx"] = page
+            image["path"] = str(path)
+            last = json.dumps(image, indent=2, ensure_ascii=False)
+    if last is not None:
+        print(last, file=args.outfile)
+    print("]", file=args.outfile)
+
+
+def main(argv: Union[List[str], None] = None) -> None:
     parser = make_argparse()
-    args = parser.parse_args()
+    args = parser.parse_args(argv)
     if args.version:
         print(playa.__version__)
         return
@@ -436,21 +530,44 @@ def main() -> None:
     logging.basicConfig(level=logging.DEBUG if args.debug else logging.WARNING)
     try:
         for path in args.pdfs:
-            with playa.open(path, space="default", max_workers=args.max_workers) as doc:
-                if args.stream is not None:  # it can't be zero either though
-                    extract_stream(doc, args)
-                elif args.content_streams:
-                    extract_page_contents(doc, args)
-                elif args.catalog:
-                    extract_catalog(doc, args)
-                elif args.text_objects:
-                    extract_text_objects(doc, args)
-                elif args.text:
-                    extract_text(doc, args)
-                elif args.structure:
-                    extract_structure(doc, args)
-                else:
-                    extract_metadata(doc, args)
+            try:
+                doc = playa.open(
+                    path,
+                    space="default",
+                    max_workers=args.max_workers,
+                    password=args.password,
+                )
+            except PDFPasswordIncorrect:
+                if args.non_interactive:
+                    raise
+                password = getpass.getpass(prompt=f"Password for {path}: ")
+                doc = playa.open(
+                    path,
+                    space="default",
+                    max_workers=args.max_workers,
+                    password=password,
+                )
+            if args.stream is not None:  # it can't be zero either though
+                extract_stream(doc, args)
+            elif args.content_streams:
+                extract_page_contents(doc, args)
+            elif args.catalog:
+                extract_catalog(doc, args)
+            elif args.content_objects:
+                extract_content_objects(doc, args)
+            elif args.text_objects:
+                extract_text_objects(doc, args)
+            elif args.text:
+                extract_text(doc, args)
+            elif args.structure:
+                extract_structure(doc, args)
+            elif args.outline:
+                extract_outline(doc, args)
+            elif args.images:
+                extract_images(doc, args)
+            else:
+                extract_metadata(doc, args)
+            doc.close()
     except RuntimeError as e:
         parser.error(f"Something went wrong:\n{e}")