+++ playa/page.py
@@ -28,16 +28,14 @@ from typing import (
 )
 
 from playa.color import (
-    PREDEFINED_COLORSPACE,
-    LITERAL_RELATIVE_COLORIMETRIC,
     BASIC_BLACK,
+    LITERAL_RELATIVE_COLORIMETRIC,
+    PREDEFINED_COLORSPACE,
     Color,
     ColorSpace,
     get_colorspace,
 )
-from playa.exceptions import (
-    PDFSyntaxError,
-)
+from playa.exceptions import PDFSyntaxError
 from playa.font import Font
 
 # FIXME: PDFObject needs to go in pdftypes somehow
@@ -56,6 +54,7 @@ from playa.pdftypes import (
     resolve1,
     stream_value,
 )
+from playa.structtree import StructTree
 from playa.utils import (
     MATRIX_IDENTITY,
     Matrix,
@@ -68,8 +67,7 @@ from playa.utils import (
     mult_matrix,
     normalize_rect,
 )
-from playa.structtree import StructTree
-from playa.worker import _deref_document, _ref_document, _ref_page, _deref_page, PageRef
+from playa.worker import PageRef, _deref_document, _deref_page, _ref_document, _ref_page
 
 if TYPE_CHECKING:
     from playa.document import Document
@@ -246,12 +244,10 @@ class Page:
             except (TypeError, ValueError):
                 log.warning("Invalid Rect in annotation: %r", annot)
                 continue
-            contents = resolve1(annot.get("Contents"))
             yield Annotation(
                 _pageref=self.pageref,
                 subtype=literal_name(subtype),
                 rect=rect,
-                contents=contents,
                 props=annot,
             )
 
@@ -291,26 +287,17 @@ class Page:
     @property
     def paths(self) -> Iterator["PathObject"]:
         """Iterator over lazy path objects."""
-        return cast(
-            Iterator["PathObject"],
-            iter(LazyInterpreter(self, self._contents, filter_class=PathObject)),
-        )
+        return self.flatten(PathObject)
 
     @property
     def images(self) -> Iterator["ImageObject"]:
         """Iterator over lazy image objects."""
-        return cast(
-            Iterator["ImageObject"],
-            iter(LazyInterpreter(self, self._contents, filter_class=ImageObject)),
-        )
+        return self.flatten(ImageObject)
 
     @property
     def texts(self) -> Iterator["TextObject"]:
         """Iterator over lazy text objects."""
-        return cast(
-            Iterator["TextObject"],
-            iter(LazyInterpreter(self, self._contents, filter_class=TextObject)),
-        )
+        return self.flatten(TextObject)
 
     @property
     def xobjects(self) -> Iterator["XObjectObject"]:
@@ -466,7 +453,6 @@ class Annotation:
       subtype: Type of annotation.
       rect: Annotation rectangle (location on page) in *default user space*
       bbox: Annotation rectangle in *device space*
-      contents: Text contents.
       props: Annotation dictionary containing all other properties
              (PDF 1.7 sec. 12.5.2).
     """
@@ -474,13 +460,42 @@ class Annotation:
     _pageref: PageRef
     subtype: str
     rect: Rect
-    contents: Union[str, None]
     props: Dict[str, PDFObject]
 
     @property
     def page(self) -> Page:
         """Containing page for this annotation."""
         return _deref_page(self._pageref)
+
+    @property
+    def contents(self) -> Union[str, None]:
+        """Text contents of annotation."""
+        contents = resolve1(self.props.get("Contents"))
+        if contents is None:
+            return None
+        return decode_text(contents)
+
+    @property
+    def name(self) -> Union[str, None]:
+        """Annotation name, uniquely identifying this annotation."""
+        name = resolve1(self.props.get("NM"))
+        if name is None:
+            return None
+        return decode_text(name)
+
+    @property
+    def mtime(self) -> Union[str, None]:
+        """String describing date and time when annotation was most recently
+        modified.
+
+        The date *should* be in the format `D:YYYYMMDDHHmmSSOHH'mm`
+        but this is in no way required (and unlikely to be implemented
+        consistently, if history is any guide).
+        """
+        mtime = resolve1(self.props.get("M"))
+        if mtime is None:
+            return None
+        return decode_text(mtime)
 
 
 @dataclass
@@ -499,9 +514,7 @@ class TextState:
         not correspond to an actual line because PDF is a presentation
         format.
       glyph_offset: The offset of the current glyph with relation to
-        the line matrix (in user space).  To get this in device space
-        you may use `playa.utils.apply_matrix_norm` with
-        `TextObject.ctm`.
+        the line matrix, in text space units.
       font: The current font.
       fontsize: The current font size, **in text space units**.
         This is often just 1.0 as it relies on the text matrix (you
@@ -562,6 +575,9 @@ class DashPattern(NamedTuple):
             return f"{self.dash} {self.phase}"
 
 
+SOLID_LINE = DashPattern((), 0)
+
+
 @dataclass
 class GraphicState:
     """PDF Graphics state (PDF 1.7 section 8.4)
@@ -578,23 +594,19 @@ class GraphicState:
       scolor: Colour used for stroking operations
       scs: Colour space used for stroking operations
       ncolor: Colour used for non-stroking operations
-      scs: Colour space used for non-stroking operations
+      ncs: Colour space used for non-stroking operations
     """
 
-    linewidth: float = 0
+    linewidth: float = 1
     linecap: int = 0
     linejoin: int = 0
     miterlimit: float = 10
-    dash: DashPattern = DashPattern((), 0)
+    dash: DashPattern = SOLID_LINE
     intent: PSLiteral = LITERAL_RELATIVE_COLORIMETRIC
     flatness: float = 1
-    # stroking color
     scolor: Color = BASIC_BLACK
-    # stroking color space
     scs: ColorSpace = PREDEFINED_COLORSPACE["DeviceGray"]
-    # non stroking color
     ncolor: Color = BASIC_BLACK
-    # non stroking color space
     ncs: ColorSpace = PREDEFINED_COLORSPACE["DeviceGray"]
 
 
@@ -741,7 +753,7 @@ class TagObject(ContentObject):
         return 0
 
     @property
-    def mcs(self) -> Union[MarkedContent, None]:
+    def mcs(self) -> MarkedContent:
         """The marked content tag for this object."""
         return self._mcs
 
@@ -984,6 +996,8 @@ class GlyphObject(ContentObject):
       matrix: rendering matrix for this glyph, which transforms text
               space (*not glyph space!*) coordinates to device space.
       bbox: glyph bounding box in device space.
+      text_space_bbox: glyph bounding box in text space (i.e. before
+                       any possible coordinate transformation)
       corners: Is the transformed bounding box rotated or skewed such
                that all four corners need to be calculated (derived
                from matrix but precomputed for speed)
@@ -1003,24 +1017,7 @@ class GlyphObject(ContentObject):
 
     @property
     def bbox(self) -> Rect:
-        tstate = self.textstate
-        font = tstate.font
-        assert font is not None
-        if font.vertical:
-            textdisp = font.char_disp(self.cid)
-            assert isinstance(textdisp, tuple)
-            (vx, vy) = textdisp
-            if vx is None:
-                vx = tstate.fontsize * 0.5
-            else:
-                vx = vx * tstate.fontsize * 0.001
-            vy = (1000 - vy) * tstate.fontsize * 0.001
-            x0, y0 = (-vx, vy + tstate.rise + self.adv)
-            x1, y1 = (-vx + tstate.fontsize, vy + tstate.rise)
-        else:
-            x0, y0 = (0, tstate.descent + tstate.rise)
-            x1, y1 = (self.adv, tstate.descent + tstate.rise + tstate.fontsize)
-
+        x0, y0, x1, y1 = self.text_space_bbox
         if self.corners:
             return get_bound(
                 (
@@ -1039,25 +1036,50 @@ class GlyphObject(ContentObject):
                 y0, y1 = y1, y0
             return (x0, y0, x1, y1)
 
+    @property
+    def text_space_bbox(self):
+        tstate = self.textstate
+        font = tstate.font
+        assert font is not None
+        if font.vertical:
+            textdisp = font.char_disp(self.cid)
+            assert isinstance(textdisp, tuple)
+            (vx, vy) = textdisp
+            if vx is None:
+                vx = tstate.fontsize * 0.5
+            else:
+                vx = vx * tstate.fontsize * 0.001
+            vy = (1000 - vy) * tstate.fontsize * 0.001
+            x0, y0 = (-vx, vy + tstate.rise + self.adv)
+            x1, y1 = (-vx + tstate.fontsize, vy + tstate.rise)
+        else:
+            x0, y0 = (0, tstate.descent + tstate.rise)
+            x1, y1 = (self.adv, tstate.descent + tstate.rise + tstate.fontsize)
+        return (x0, y0, x1, y1)
+
 
 @dataclass
 class TextObject(ContentObject):
     """Text object (contains one or more glyphs).
 
     Attributes:
-      textstate: Text state for this object.  This is a **mutable**
-        object and you should not expect it to be valid outside the
-        context of iteration over the parent `TextObject`.
+      textstate: Text state for this object.
       args: Strings or position adjustments
+      bbox: Text bounding box in device space.
+      text_space_bbox: Text bounding box in text space (i.e. before
+                       any possible coordinate transformation)
     """
 
     textstate: TextState
     args: List[Union[bytes, float]]
     _chars: Union[List[str], None] = None
+    _bbox: Union[Rect, None] = None
+    _text_space_bbox: Union[Rect, None] = None
+    _next_tstate: Union[TextState, None] = None
 
     def __iter__(self) -> Iterator[GlyphObject]:
         """Generate glyphs for this text object"""
-        tstate = self.textstate
+        tstate = copy(self.textstate)
         font = tstate.font
         # If no font is set, we cannot do anything, since even calling
         # TJ with a displacement and no text effects requires us at
@@ -1067,6 +1089,7 @@ class TextObject(ContentObject):
                 "No font is set, will not update text state or output text: %r TJ",
                 self.args,
             )
+            self._next_tstate = tstate
             return
         assert self.ctm is not None
         # Extract all the elements so we can translate efficiently
@@ -1082,7 +1105,7 @@ class TextObject(ContentObject):
             wordspace = 0
         (x, y) = tstate.glyph_offset
         pos = y if vert else x
-        needcharspace = False
+        needcharspace = False  # Only for first glyph
         for obj in self.args:
             if isinstance(obj, (int, float)):
                 dxscale = 0.001 * tstate.fontsize * scaling
@@ -1092,10 +1115,9 @@ class TextObject(ContentObject):
                 for cid, text in font.decode(obj):
                     if needcharspace:
                         pos += charspace
-                    tstate.glyph_offset = (x, pos) if vert else (pos, y)
                     textwidth = font.char_width(cid)
                     adv = textwidth * tstate.fontsize * scaling
-                    x, y = tstate.glyph_offset
+                    x, y = tstate.glyph_offset = (x, pos) if vert else (pos, y)
                     glyph = GlyphObject(
                         _pageref=self._pageref,
                         gstate=self.gstate,
@@ -1115,6 +1137,101 @@ class TextObject(ContentObject):
                         pos += wordspace
                     needcharspace = True
         tstate.glyph_offset = (x, pos) if vert else (pos, y)
+        if self._next_tstate is None:
+            self._next_tstate = tstate
+
+    @property
+    def text_space_bbox(self):
+        if self._text_space_bbox is not None:
+            return self._text_space_bbox
+        # No need to save tstate as we do not update it below
+        tstate = self.textstate
+        font = tstate.font
+        if font is None:
+            log.warning(
+                "No font is set, will not update text state or output text: %r TJ",
+                self.args,
+            )
+            self._text_space_bbox = BBOX_NONE
+            self._next_tstate = tstate
+            return self._text_space_bbox
+        if len(self.args) == 0:
+            self._text_space_bbox = BBOX_NONE
+            self._next_tstate = tstate
+            return self._text_space_bbox
+        scaling = tstate.scaling * 0.01
+        charspace = tstate.charspace * scaling
+        wordspace = tstate.wordspace * scaling
+        vert = font.vertical
+        if font.multibyte:
+            wordspace = 0
+        (x, y) = tstate.glyph_offset
+        pos = y if vert else x
+        needcharspace = False  # Only for first glyph
+        if vert:
+            x0 = x1 = x
+            y0 = y1 = y
+        else:
+            # These do not change!
+            x0 = x1 = x
+            y0 = y + tstate.descent + tstate.rise
+            y1 = y0 + tstate.fontsize
+        for obj in self.args:
+            if isinstance(obj, (int, float)):
+                dxscale = 0.001 * tstate.fontsize * scaling
+                pos -= obj * dxscale
+                needcharspace = True
+            else:
+                for cid, _ in font.decode(obj):
+                    if needcharspace:
+                        pos += charspace
+                    textwidth = font.char_width(cid)
+                    adv = textwidth * tstate.fontsize * scaling
+                    x, y = (x, pos) if vert else (pos, y)
+                    if vert:
+                        textdisp = font.char_disp(cid)
+                        assert isinstance(textdisp, tuple)
+                        (vx, vy) = textdisp
+                        if vx is None:
+                            vx = tstate.fontsize * 0.5
+                        else:
+                            vx = vx * tstate.fontsize * 0.001
+                        vy = (1000 - vy) * tstate.fontsize * 0.001
+                        x0 = min(x0, x - vx)
+                        y0 = min(y0, y + vy + tstate.rise + adv)
+                        x1 = max(x1, x - vx + tstate.fontsize)
+                        y1 = max(y1, y + vy + tstate.rise)
+                    else:
+                        x1 = x + adv
+                    pos += adv
+                    if cid == 32 and wordspace:
+                        pos += wordspace
+                    needcharspace = True
+        if self._next_tstate is None:
+            self._next_tstate = copy(tstate)
+            self._next_tstate.glyph_offset = (x, pos) if vert else (pos, y)
+        self._text_space_bbox = (x0, y0, x1, y1)
+        return self._text_space_bbox
+
+    @property
+    def next_textstate(self) -> TextState:
+        if self._next_tstate is not None:
+            return self._next_tstate
+        _ = self.text_space_bbox
+        assert self._next_tstate is not None
+        return self._next_tstate
+
+    @property
+    def bbox(self) -> Rect:
+        # We specialize this to avoid it having side effects on the
+        # text state (already it's a bit of a footgun that __iter__
+        # does that...), but also because we know all glyphs have the
+        # same text matrix and thus we can avoid a lot of multiply
+        if self._bbox is not None:
+            return self._bbox
+        matrix = mult_matrix(self.textstate.line_matrix, self.ctm)
+        self._bbox = get_transformed_bound(matrix, self.text_space_bbox)
+        return self._bbox
 
     @property
     def chars(self) -> str:
@@ -1301,6 +1418,8 @@ class LazyInterpreter:
                         co = method()
                     if co is not None:
                         yield co
+                    if isinstance(co, TextObject):
+                        self.textstate = co.next_textstate
                 else:
                     # TODO: This can get very verbose
                     log.warning("Unknown operator: %r", obj)
@@ -1424,13 +1543,12 @@ class LazyInterpreter:
                     "Ignoring non-string/number %r in text object %r", s, strings
                 )
         obj = self.create(TextObject, textstate=self.textstate, args=args)
-        if has_text:
-            return obj
-        else:
+        if obj is not None:
+            if has_text:
+                return obj
             # Even without text, TJ can still update the line matrix (ugh!)
-            if obj is not None:
-                for _ in obj:
-                    pass
+            assert isinstance(obj, TextObject)
+            self.textstate = obj.next_textstate
         return None
 
     def do_Tj(self, s: PDFObject) -> Union[ContentObject, None]: