+++ playa/utils.py
@@ -10,7 +10,6 @@ from typing import (
     Union,
 )
 
-
 # from sys import maxint as INF doesn't work anymore under Python3, but PDF
 # still uses 32 bits ints
 INF = (1 << 31) - 1