+++ tests/test_lazy_api.py
@@ -11,9 +11,9 @@ import pytest
 import playa
 from playa.color import PREDEFINED_COLORSPACE, Color
 from playa.exceptions import PDFEncryptionError
-from playa.utils import get_transformed_bound, get_bound, apply_matrix_pt, Matrix
+from playa.utils import Matrix, apply_matrix_pt, get_bound, get_transformed_bound
 
-from .data import TESTDIR, ALLPDFS, PASSWORDS, XFAILS, CONTRIB
+from .data import ALLPDFS, CONTRIB, PASSWORDS, TESTDIR, XFAILS
 
 
 @pytest.mark.skipif(not CONTRIB.exists(), reason="contrib samples not present")
@@ -62,7 +62,12 @@ def test_open_lazy(path: Path) -> None:
             with playa.open(path, password=password) as doc:
                 for page in doc.pages:
                     for obj in page:
-                        beach.append((obj.object_type, obj.bbox))
+                        try:
+                            beach.append((obj.object_type, obj.bbox))
+                        except ValueError as e:
+                            if "not enough values" in str(e):
+                                continue
+                            raise e
         except PDFEncryptionError:
             pytest.skip("cryptography package not installed")
 
@@ -99,6 +104,22 @@ def test_rotated_glyphs() -> None:
                     width = x1 - x0
                     assert width > 6
         assert "".join(chars) == "R18,00"
+
+
+def test_rotated_text_objects() -> None:
+    """Verify specializations of bbox for text."""
+    with playa.open(TESTDIR / "rotated.pdf") as pdf:
+        # Ensure that the text bbox is the same as the bounds of the
+        # glyph bboxes (this will also ensure no side effects)
+        for text in pdf.pages[0].texts:
+            bbox = text.bbox
+            points = []
+            for glyph in text:
+                x0, y0, x1, y1 = glyph.bbox
+                print(glyph.text, ":", glyph.bbox)
+                points.append((x0, y0))
+                points.append((x1, y1))
+            assert bbox == pytest.approx(get_bound(points))
 
 
 def test_rotated_bboxes() -> None: